Fixed request changes
- Cleared goMain.ts, moved logic to lintCurrentPackage and lintWorkspace in goLint.ts
- Moved function handleDiagnosticErrors from goMain.ts to util.ts
- Fixed tests
