+++ src/util.ts
@@ -11,6 +11,7 @@ import TelemetryReporter from 'vscode-extension-telemetry';
 import fs = require('fs');
 import os = require('os');
 import { outputChannel } from './goStatus';
+import { errorDiagnosticCollection, warningDiagnosticCollection } from './goMain';
 
 const extensionId: string = 'lukehoban.Go';
 const extensionVersion: string = vscode.extensions.getExtension(extensionId).packageJSON.version;
@@ -565,4 +566,47 @@ export function runTool(args: string[], cwd: string, severity: string, useStdErr
 			}
 		});
 	});
+}
+
+export function handleDiagnosticErrors(document: vscode.TextDocument) {
+	return (errors: ICheckResult[]) => {
+		let diagnosticMap: Map<string, Map<vscode.DiagnosticSeverity, vscode.Diagnostic[]>> = new Map();
+
+		errors.forEach(error => {
+			let canonicalFile = vscode.Uri.file(error.file).toString();
+			let startColumn = 0;
+			let endColumn = 1;
+			if (document && document.uri.toString() === canonicalFile) {
+				let range = new vscode.Range(error.line - 1, 0, error.line - 1, document.lineAt(error.line - 1).range.end.character + 1);
+				let text = document.getText(range);
+				let [_, leading, trailing] = /^(\s*).*(\s*)$/.exec(text);
+				startColumn = leading.length;
+				endColumn = text.length - trailing.length;
+			}
+			let range = new vscode.Range(error.line - 1, startColumn, error.line - 1, endColumn);
+			let severity = mapSeverityToVSCodeSeverity(error.severity);
+			let diagnostic = new vscode.Diagnostic(range, error.msg, severity);
+			let diagnostics = diagnosticMap.get(canonicalFile);
+			if (!diagnostics) {
+				diagnostics = new Map<vscode.DiagnosticSeverity, vscode.Diagnostic[]>();
+			}
+			if (!diagnostics[severity]) {
+				diagnostics[severity] = [];
+			}
+			diagnostics[severity].push(diagnostic);
+			diagnosticMap.set(canonicalFile, diagnostics);
+		});
+		diagnosticMap.forEach((diagMap, file) => {
+			errorDiagnosticCollection.set(vscode.Uri.parse(file), diagMap[vscode.DiagnosticSeverity.Error]);
+			warningDiagnosticCollection.set(vscode.Uri.parse(file), diagMap[vscode.DiagnosticSeverity.Warning]);
+		});
+	};
+}
+
+function mapSeverityToVSCodeSeverity(sev: string): vscode.DiagnosticSeverity {
+	switch (sev) {
+		case 'error': return vscode.DiagnosticSeverity.Error;
+		case 'warning': return vscode.DiagnosticSeverity.Warning;
+		default: return vscode.DiagnosticSeverity.Error;
+	}
 }
